fix(sparse): clear the --all-targets clippy gate for fsci-sparse (nwx8m cause 1)
frankenscipy-nwx8m cause 1. `cargo clippy -p fsci-sparse --all-targets -- -D
warnings` now FINISHES CLEAN; it previously failed with 8 errors across three
targets. Cause 2 (--all-features unresolved imports) is untouched and the bead
stays open -- see below.

WHAT WAS FIXED, and why each fix is the one it is:

  perf_spsolve.rs -- `spsolve_triangular` was imported ungated but only used by
  #[cfg(feature = "sparse-incumbent-bench")] code, so the import is dead under
  default features. Gated the import to match its only use site rather than
  deleting it.

  perf_spsolve.rs -- laplacian_3d_cubic / laplacian_3d_cuboid flagged
  `never used`. NOT DELETED. Their own comment says they exist for the
  profile-first tensor campaign; they are fixture specs, and deleting a spec to
  silence dead_code is exactly the mistake recorded against iit9c. Added
  #[allow(dead_code)] with the reason inline.

  profile_gmres_arnoldi.rs and construct.rs -- type_complexity satisfied with
  named `type` definitions (ArnoldiKernel, DiagsCase) rather than an allow,
  which is what the lint actually asks for.

  linalg.rs x4 and ops.rs x1 -- needless_range_loop rewritten to
  zip/enumerate. These are assertion loops, so the rewrites preserve the index
  variables the failure messages interpolate; behaviour is identical.

  linalg.rs (normalized-laplacian scaling) -- `for idx in indptr[i]..indptr[i+1]
  { let j = indices[idx]; .. }` became `for &j in &indices[indptr[i]..indptr[i+1]]`.
  This one is LIBRARY code, not test code; the transformation is exactly
  equivalent, and the crate suite is green after it.

  ops.rs (CSC reconstruction) -- kept the index form under a scoped
  #[allow(clippy::needless_range_loop)] WITH a written reason: that loop walks
  columns and scatters into got[r][c] where r comes from the indices array and
  varies per entry, so there is no row iterator to borrow and an iterator
  rewrite would obscure the access pattern. One documented suppression is
  honest; blanket-allowing the lint would not be.

  benches/sparse_bench.rs -- an unnecessary usize->usize cast, newly VISIBLE
  once the earlier errors stopped aborting the build. Worth noting: the bead
  counted 8 errors, and clearing them exposed a 9th that had been masked.

EVIDENCE: clippy -p fsci-sparse --all-targets -D warnings -> Finished, no
errors (was 8 across profile_gmres_arnoldi, perf_spsolve and lib test).
cargo test -p fsci-sparse --lib -> 426 passed, 0 failed, 4 ignored, unchanged
from before the edits, which is what makes the loop rewrites safe to believe.
cargo fmt -p fsci-sparse --check exit 0.

CAUSE 2 REMAINS OPEN, deliberately. --all-features still fails with E0432 on
GMRES_BATCH_FORCE_SEQUENTIAL, LGMRES_/QMR_BATCH_FORCE_SEQUENTIAL,
ITERATIVE_BATCH_LAST_WORKERS and the CG_* controls. Those are the same family as
frankenscipy-vacuous-perf-toggles-qcuyy: perf harnesses driving levers that were
removed. Fixing it means either restoring the implementations (which is
qcuyy/ct8n2 work, and restoring bare statics nothing reads would recreate the
vacuous-toggle defect I fixed earlier this session) or deleting harness code.
Neither is a lint fix, so it is not being done under a lint bead.

+++ crates/fsci-sparse/benches/sparse_bench.rs
@@ -747,7 +747,7 @@ fn bench_diags_validate_ab(c: &mut Criterion) {
     let offsets: Vec<isize> = (-band..=band).collect();
     let diagonals: Vec<Vec<f64>> = offsets
         .iter()
-        .map(|&o| vec![1.0; n - o.unsigned_abs() as usize])
+        .map(|&o| vec![1.0; n - o.unsigned_abs()])
         .collect();
     let shape = Some(Shape2D::new(n, n));
